fix: address pr review comments for accessibility and code quality
- ActivationCheckpointConfig: default RecomputeStrategy to None since
  Selective/Full are not yet implemented (avoids NotImplementedException)
- LoadBalancedPartitionStrategy: enforce first boundary == 0 to prevent
  orphaned parameters, add overflow checks for long-to-int casts,
  remove useless local variable assignment
- Pipeline schedules (1F1B, Interleaved, ZB-H1/H2/V, LoopedBFS): use
  double arithmetic directly in EstimateBubbleFraction to avoid
  precision loss from long-to-double implicit conversions
- PipelineParallelModel: combine nested if statements, preserve inner
  exceptions in catch clauses for better debugging

+++ src/DistributedTraining/Interleaved1F1BSchedule.cs
@@ -181,9 +181,9 @@ public class Interleaved1F1BSchedule : IPipelineSchedule
 
         // Interleaved 1F1B bubble: (P-1) / (2*M*V + P - 1)
         // V times smaller than standard 1F1B
-        long p = numStages;
-        long m = numMicroBatches;
-        long v = _virtualStagesPerRank;
-        return (double)(p - 1) / (2 * m * v + p - 1);
+        double p = numStages;
+        double m = numMicroBatches;
+        double v = _virtualStagesPerRank;
+        return (p - 1) / (2.0 * m * v + p - 1.0);
     }
 }
